Wp0625

+++ Classes/Scene/Options/HeroScene.h
@@ -4,29 +4,29 @@
 #ifndef __HERO_SCENE_H__
 #define __HERO_SCENE_H__
 
-#define HERO_BACK_ITEM_POSITION_X origin.x + heroBackItem->getContentSize().width / 2
-#define HERO_BACK_ITEM_POSITION_Y origin.y + heroBackItem->getContentSize().height / 2
+#define HERO_BACK_ITEM_POSITION_X static_cast<float>(origin.x + heroBackItem->getContentSize().width / 2)
+#define HERO_BACK_ITEM_POSITION_Y static_cast<float>(origin.y + heroBackItem->getContentSize().height / 2)
 
-#define HERO_NAME_LABEL_POSITION_X origin.x + visibleSize.width / 2
-#define HERO_NAME_LABEL_POSITION_Y origin.y + visibleSize.height - heroNameLabel->getContentSize().height
+#define HERO_NAME_LABEL_POSITION_X static_cast<float>(origin.x + visibleSize.width / 2)
+#define HERO_NAME_LABEL_POSITION_Y static_cast<float>(origin.y + visibleSize.height - heroNameLabel->getContentSize().height)
 
-#define HERO1_SELECTED_BUTTON_POSITION_X origin.x + visibleSize.width / 4
-#define HERO1_SELECTED_BUTTON_POSITION_Y origin.y + visibleSize.height / 5*4
+#define HERO1_SELECTED_BUTTON_POSITION_X static_cast<float>(origin.x + visibleSize.width / 4)
+#define HERO1_SELECTED_BUTTON_POSITION_Y static_cast<float>(origin.y + visibleSize.height / 5*4)
 
-#define HERO2_SELECTED_BUTTON_POSITION_X origin.x + visibleSize.width / 4
-#define HERO2_SELECTED_BUTTON_POSITION_Y origin.y + visibleSize.height / 5*3
+#define HERO2_SELECTED_BUTTON_POSITION_X static_cast<float>(origin.x + visibleSize.width / 4)
+#define HERO2_SELECTED_BUTTON_POSITION_Y static_cast<float>(origin.y + visibleSize.height / 5*3)
 
-#define HERO3_SELECTED_BUTTON_POSITION_X origin.x + visibleSize.width / 4
-#define HERO3_SELECTED_BUTTON_POSITION_Y origin.y + visibleSize.height / 5*2
+#define HERO3_SELECTED_BUTTON_POSITION_X static_cast<float>(origin.x + visibleSize.width / 4)
+#define HERO3_SELECTED_BUTTON_POSITION_Y static_cast<float>(origin.y + visibleSize.height / 5*2)
 
-#define HERO4_SELECTED_BUTTON_POSITION_X origin.x + visibleSize.width / 4
-#define HERO4_SELECTED_BUTTON_POSITION_Y origin.y + visibleSize.height / 5*1
+#define HERO4_SELECTED_BUTTON_POSITION_X static_cast<float>(origin.x + visibleSize.width / 4)
+#define HERO4_SELECTED_BUTTON_POSITION_Y static_cast<float>(origin.y + visibleSize.height / 5*1)
 
-#define HERO_CONFIRM_BUTTON_POSITION_X  origin.x + visibleSize.width / 4*3
-#define HERO_CONFIRM_BUTTON_POSITION_Y  origin.y + visibleSize.height / 5
+#define HERO_CONFIRM_BUTTON_POSITION_X  static_cast<float>(origin.x + visibleSize.width / 4*3)
+#define HERO_CONFIRM_BUTTON_POSITION_Y  static_cast<float>(origin.y + visibleSize.height / 5)
 
-#define HERO_INFO_ITEM_POSITION_X origin.x + visibleSize.width - heroInformation->getContentSize().width
-#define HERO_INFO_ITEM_POSITION_Y visibleSize.height - 1.2 * heroInformation->getContentSize().height / 2
+#define HERO_INFO_ITEM_POSITION_X static_cast<float>(origin.x + visibleSize.width - heroInformation->getContentSize().width)
+#define HERO_INFO_ITEM_POSITION_Y static_cast<float>(visibleSize.height - 1.2 * heroInformation->getContentSize().height / 2)
 
 #include "cocos2d.h"
 #include "Scene/MainMenu/MainMenuScene.h"
